requirements.test.txt
requests==2.28.1
pytest==7.1.2
pytest-dependency==0.5.1
pytest-selenium
pythonping